move number_literal_to_steel into an impl IntoSteelval for &NumberLiteral
