+++ crates/steel-core/src/compiler/code_gen.rs
@@ -4,7 +4,6 @@ use super::{
         Analysis,
         CallKind::{Normal, SelfTailCall, TailCall},
     },
-    program::number_literal_to_steel,
 };
 use crate::{
     compiler::passes::analysis::IdentifierStatus::{
@@ -23,6 +22,7 @@ use crate::{
         tryfrom_visitor::TryFromExprKindForSteelVal,
         visitors::VisitorMut,
     },
+    rvals::IntoSteelVal,
     stop, SteelVal,
 };
 use smallvec::SmallVec;
@@ -55,7 +55,7 @@ fn eval_atom(t: &SyntaxObject) -> Result<SteelVal> {
 fn try_eval_atom(t: &SyntaxObject) -> Option<SteelVal> {
     match &t.ty {
         TokenType::BooleanLiteral(b) => Some((*b).into()),
-        TokenType::Number(n) => number_literal_to_steel(n).ok(),
+        TokenType::Number(n) => (&**n).into_steelval().ok(),
         TokenType::StringLiteral(s) => Some(SteelVal::StringV(s.to_string().into())),
         TokenType::CharacterLiteral(c) => Some(SteelVal::CharV(*c)),
         // TODO: Keywords shouldn't be misused as an expression - only in function calls are keywords allowed
@@ -69,7 +69,7 @@ fn try_eval_atom(t: &SyntaxObject) -> Option<SteelVal> {
 fn try_eval_atom_with_context(t: &SyntaxObject) -> Result<SteelVal> {
     match &t.ty {
         TokenType::BooleanLiteral(b) => Ok((*b).into()),
-        TokenType::Number(n) => number_literal_to_steel(n).map_err(|e| e.with_span(t.span)),
+        TokenType::Number(n) => (&**n).into_steelval().map_err(|e| e.with_span(t.span)),
         TokenType::StringLiteral(s) => Ok(SteelVal::StringV(s.clone().into())),
         TokenType::CharacterLiteral(c) => Ok(SteelVal::CharV(*c)),
         TokenType::Keyword(k) => Ok(SteelVal::SymbolV(k.clone().into())),
