+++ crates/steel-core/src/parser/parser.rs
@@ -2,6 +2,7 @@ use crate::primitives::numbers::make_polar;
 use crate::rvals::{IntoSteelVal, SteelComplex, SteelString};
 use crate::{parser::tokens::TokenType::*, rvals::FromSteelVal};
 
+use num_bigint::BigInt;
 use num_rational::{BigRational, Rational32};
 use std::borrow::Cow;
 use std::str;
@@ -176,6 +177,52 @@ impl IntoSteelVal for NumberLiteral {
             NumberLiteral::Polar(r, theta) => {
                 let r = real_literal_to_steelval(r)?;
                 let theta = real_literal_to_steelval(theta)?;
+
+                make_polar(&r, &theta)
+            }
+        }
+    }
+}
+
+impl<'a> IntoSteelVal for &'a NumberLiteral {
+    fn into_steelval(self) -> Result<SteelVal, SteelErr> {
+        // real_to_steel does some cloning of bignums. It may be possible to optimize this away.
+        let real_to_steel = |re: &RealLiteral| match re {
+            RealLiteral::Int(IntLiteral::Small(i)) => i.into_steelval(),
+            RealLiteral::Int(IntLiteral::Big(i)) => i.clone().into_steelval(),
+            RealLiteral::Rational(n, d) => match (n, d) {
+                (IntLiteral::Small(n), IntLiteral::Small(d)) => {
+                    match (i32::try_from(*n), i32::try_from(*d)) {
+                        (Ok(n), Ok(0)) => {
+                            stop!(BadSyntax => format!("division by zero in {:?}/0", n))
+                        }
+                        (Ok(n), Ok(d)) => Rational32::new(n, d).into_steelval(),
+                        _ => BigRational::new(BigInt::from(*n), BigInt::from(*d)).into_steelval(),
+                    }
+                }
+                (IntLiteral::Small(n), IntLiteral::Big(d)) => {
+                    BigRational::new(BigInt::from(*n), *d.clone()).into_steelval()
+                }
+                (IntLiteral::Big(n), IntLiteral::Small(d)) => {
+                    BigRational::new(*n.clone(), BigInt::from(*d)).into_steelval()
+                }
+                (IntLiteral::Big(n), IntLiteral::Big(d)) => {
+                    BigRational::new(*n.clone(), *d.clone()).into_steelval()
+                }
+            },
+            RealLiteral::Float(f) => f.into_steelval(),
+        };
+
+        match self {
+            NumberLiteral::Real(re) => real_to_steel(re),
+            NumberLiteral::Complex(re, im) => SteelComplex {
+                re: real_to_steel(re)?,
+                im: real_to_steel(im)?,
+            }
+            .into_steelval(),
+            NumberLiteral::Polar(r, theta) => {
+                let r = real_to_steel(r)?;
+                let theta = real_to_steel(theta)?;
 
                 make_polar(&r, &theta)
             }
